+++ src/main/java/br/edu/ifrs/palestra/dto/DisciplinaDTO.java
@@ -1,4 +1,4 @@
 package br.edu.ifrs.palestra.dto;
 
-public record DisciplinaDTO(String nome, Integer semestre) {
+public record DisciplinaDTO(String nome, Long semestre) {
 }
