+++ src/main/java/br/edu/ifrs/palestra/dto/TurmaDTO.java
@@ -1,4 +1,4 @@
 package br.edu.ifrs.palestra.dto;
 
-public record TurmaDTO(String nome, Integer disciplina_id) {
+public record TurmaDTO(String nome, Long disciplina_id) {
 }
