+++ src/main/java/br/edu/ifrs/palestra/service/interfaces/TurmaService.java
@@ -4,7 +4,7 @@ import br.edu.ifrs.palestra.dto.TurmaDTO;
 import br.edu.ifrs.palestra.model.Turma;
 
 public interface TurmaService extends CrudService<Turma, TurmaDTO> {
-    Turma matricular(int turma_id, int aluno_id);
+    Turma matricular(Long turma_id, Long aluno_id);
 
-    Turma cancelarMatricula(int turma_id, int aluno_id);
+    Turma cancelarMatricula(Long turma_id, Long aluno_id);
 }
